wjx/kyle-robot-toolbox/requirements.txt
setuptools==58.2.0
jupyter
numpy
numpy-quaternion
scipy
matplotlib
seaborn
opencv-python
opencv-contrib-python
torch
torchvision
torchaudio
onnx
ultralytics
open3d
trimesh
PyYaml
absl-py
pybind11
